+++ script.js
@@ -1,1 +1,3 @@
+
+
 const task = document.getElementById('task');
